Re-organize TS api directory, rename some functions, add more helpers
