+++ src/typescript/api/src/emojicoin_dot_fun/consts.ts
@@ -5,3 +5,7 @@ export const MAX_GAS_FOR_PUBLISH = 1500000;
 export const COIN_FACTORY_MODULE_NAME = "coin_factory";
 export const EMOJICOIN_DOT_FUN_MODULE_NAME = "emojicoin_dot_fun";
 export const MODULE_ADDRESS = (() => AccountAddress.from(process.env.MODULE_ADDRESS!))();
+export const DEFAULT_REGISTER_MARKET_GAS_OPTIONS = {
+  maxGasAmount: ONE_APT / 100,
+  gasUnitPrice: 100,
+};
