+++ src/typescript/api/src/emojicoin_dot_fun/utils.ts
@@ -4,15 +4,11 @@ import {
   Aptos,
   Hex,
   type HexInput,
-  type Uint64,
-  type Uint128,
   DeriveScheme,
-  type TypeTag,
-  parseTypeTag,
   type AptosConfig,
 } from "@aptos-labs/ts-sdk";
 import { sha3_256 } from "@noble/hashes/sha3";
-import { COIN_FACTORY_MODULE_NAME, EMOJICOIN_DOT_FUN_MODULE_NAME, MODULE_ADDRESS } from "./consts";
+import { EMOJICOIN_DOT_FUN_MODULE_NAME } from "./consts";
 
 /**
  * Sleep the current thread for the given amount of time
@@ -22,6 +18,10 @@ export async function sleep(timeMs: number): Promise<null> {
   return new Promise((resolve) => {
     setTimeout(resolve, timeMs);
   });
+}
+
+export function toAptos(aptos: Aptos | AptosConfig): Aptos {
+  return aptos instanceof Aptos ? aptos : new Aptos(aptos);
 }
 
 export function toConfig(aptos: Aptos | AptosConfig): AptosConfig {
@@ -75,125 +75,3 @@ export async function getRegistryAddress(args: {
   });
   return registryAddressResource.registry_address;
 }
-
-/**
- * Get the derived market address and TypeTags for the given registry address and symbol bytes.
- *
- * @param registryAddress The contract's registry address.
- * @param symbolBytes The emojicoin's full symbol bytes.
- * @returns The derived market address and TypeTags.
- */
-export function getEmojicoinMarketAddressAndTypeTags(args: {
-  registryAddress: AccountAddressInput;
-  symbolBytes: HexInput;
-}): [AccountAddress, TypeTag, TypeTag] {
-  const registryAddress = AccountAddress.from(args.registryAddress);
-  const symbolBytes = Hex.fromHexInput(args.symbolBytes);
-  const marketAddress = deriveEmojicoinPublisherAddress({
-    registryAddress,
-    emojis: [symbolBytes.toStringWithoutPrefix()],
-  });
-
-  return [
-    marketAddress,
-    parseTypeTag(`${marketAddress.toString()}::${COIN_FACTORY_MODULE_NAME}::Emojicoin`),
-    parseTypeTag(`${marketAddress.toString()}::${COIN_FACTORY_MODULE_NAME}::EmojicoinLP`),
-  ];
-}
-
-export type MarketMetadata = {
-  market_id: Uint64;
-  market_address: AccountAddress;
-  emoji_bytes: Hex;
-};
-
-export type SequenceInfo = {
-  nonce: Uint64;
-  last_bump_time: Uint64;
-};
-
-export type MarketResource = {
-  metadata: MarketMetadata;
-  sequence_info: SequenceInfo;
-  extend_ref: ExtendRef;
-  clamm_virtual_reserves: Reserves;
-  cpamm_real_reserves: Reserves;
-  lp_coin_supply: Uint128;
-};
-
-export type ExtendRef = {
-  self: AccountAddress;
-};
-
-export type Reserves = {
-  base: Uint64;
-  quote: Uint64;
-};
-
-export async function getMarketResource(args: {
-  aptos: Aptos;
-  moduleAddress: AccountAddressInput;
-  objectAddress: AccountAddressInput;
-}): Promise<MarketResource> {
-  const { aptos } = args;
-  const moduleAddress = AccountAddress.from(args.moduleAddress);
-  const objectAddress = AccountAddress.from(args.objectAddress);
-  const marketResource = await aptos.getAccountResource({
-    accountAddress: objectAddress,
-    resourceType: `${moduleAddress.toString()}::${EMOJICOIN_DOT_FUN_MODULE_NAME}::Market`,
-  });
-  return {
-    metadata: {
-      market_id: BigInt(marketResource.metadata.market_id),
-      market_address: AccountAddress.from(marketResource.metadata.market_address),
-      emoji_bytes: Hex.fromHexString(marketResource.metadata.emoji_bytes),
-    },
-    sequence_info: {
-      nonce: BigInt(marketResource.sequence_info.nonce),
-      last_bump_time: BigInt(marketResource.sequence_info.last_bump_time),
-    },
-    extend_ref: {
-      self: AccountAddress.from(marketResource.extend_ref.self),
-    },
-    clamm_virtual_reserves: {
-      base: BigInt(marketResource.clamm_virtual_reserves.base),
-      quote: BigInt(marketResource.clamm_virtual_reserves.quote),
-    },
-    cpamm_real_reserves: {
-      base: BigInt(marketResource.cpamm_real_reserves.base),
-      quote: BigInt(marketResource.cpamm_real_reserves.quote),
-    },
-    lp_coin_supply: BigInt(marketResource.lp_coin_supply),
-  };
-}
-
-export type ChatEvent = {
-  market_metadata: MarketMetadata;
-  emit_time: Uint64;
-  emit_market_nonce: Uint64;
-  user: AccountAddress;
-  message: string;
-  user_emojicoin_balance: Uint64;
-  circulating_supply: Uint64;
-  balance_as_fraction_of_circulating_supply_q64: Uint128;
-};
-
-export const chatEventTypeTag = (): TypeTag =>
-  parseTypeTag(`${MODULE_ADDRESS.toString()}::${EMOJICOIN_DOT_FUN_MODULE_NAME}::Chat`);
-
-export const parseChatEvent = (data: any): ChatEvent => ({
-  market_metadata: {
-    market_id: BigInt(data.market_metadata.market_id),
-    market_address: AccountAddress.from(data.market_metadata.market_address),
-    emoji_bytes: Hex.fromHexString(data.market_metadata.emoji_bytes),
-  },
-  emit_time: BigInt(data.emit_time),
-  emit_market_nonce: BigInt(data.emit_market_nonce),
-  user: data.user,
-  message: data.message,
-  user_emojicoin_balance: BigInt(data.user_emojicoin_balance),
-  circulating_supply: BigInt(data.circulating_supply),
-  balance_as_fraction_of_circulating_supply_q64: BigInt(
-    data.balance_as_fraction_of_circulating_supply_q64
-  ),
-});
