+++ src/typescript/api/src/markets/utils.ts
@@ -1,0 +1,78 @@
+import {
+  type Account,
+  AccountAddress,
+  type AccountAddressInput,
+  type Aptos,
+  type AptosConfig,
+  Hex,
+  type HexInput,
+  type InputGenerateTransactionOptions,
+  parseTypeTag,
+} from "@aptos-labs/ts-sdk";
+import {
+  toConfig,
+  DEFAULT_REGISTER_MARKET_GAS_OPTIONS,
+  EmojicoinDotFun,
+  COIN_FACTORY_MODULE_NAME,
+  deriveEmojicoinPublisherAddress,
+} from "../emojicoin_dot_fun";
+import { type EmojicoinInfo } from "../types/contract";
+
+/**
+ * Get the derived market address and TypeTags for the given registry address and symbol bytes.
+ *
+ * @param registryAddress The contract's registry address.
+ * @param symbolBytes The emojicoin's full symbol bytes.
+ * @returns The derived market address and TypeTags.
+ */
+export function getEmojicoinMarketAddressAndTypeTags(args: {
+  registryAddress: AccountAddressInput;
+  symbolBytes: HexInput;
+}): EmojicoinInfo {
+  const registryAddress = AccountAddress.from(args.registryAddress);
+  const symbolBytes = Hex.fromHexInput(args.symbolBytes);
+  const marketAddress = deriveEmojicoinPublisherAddress({
+    registryAddress,
+    emojis: [symbolBytes.toStringWithoutPrefix()],
+  });
+
+  return {
+    marketAddress,
+    emojicoin: parseTypeTag(`${marketAddress.toString()}::${COIN_FACTORY_MODULE_NAME}::Emojicoin`),
+    emojicoinLP: parseTypeTag(
+      `${marketAddress.toString()}::${COIN_FACTORY_MODULE_NAME}::EmojicoinLP`
+    ),
+  };
+}
+
+export const registerMarketAndGetEmojicoinInfo = async (args: {
+  aptos: Aptos | AptosConfig;
+  registryAddress: AccountAddressInput;
+  emojis: Array<HexInput>;
+  sender: Account;
+  integrator: AccountAddressInput;
+  options?: InputGenerateTransactionOptions;
+}): Promise<EmojicoinInfo> => {
+  const { aptos, emojis, sender, integrator } = args;
+  const aptosConfig = toConfig(aptos);
+  const options = args.options || DEFAULT_REGISTER_MARKET_GAS_OPTIONS;
+  const res = await EmojicoinDotFun.RegisterMarket.submit({
+    aptosConfig,
+    registrant: sender,
+    emojis,
+    integrator,
+    options,
+  });
+
+  if (!res.success) {
+    throw new Error(`Failed to register market: ${res.vm_status}, \nHash: ${res.hash}`);
+  }
+
+  const registryAddress = AccountAddress.from(args.registryAddress);
+  const { marketAddress, emojicoin, emojicoinLP } = getEmojicoinMarketAddressAndTypeTags({
+    registryAddress,
+    symbolBytes: emojis.map((v) => Hex.fromHexInput(v).toStringWithoutPrefix()).join(""),
+  });
+
+  return { marketAddress, emojicoin, emojicoinLP };
+};
