+++ src/typescript/api/src/types/core.ts
@@ -1,0 +1,10 @@
+import { type AccountAddress, type Uint64 } from "@aptos-labs/ts-sdk";
+
+export type SequenceInfo = {
+  nonce: Uint64;
+  last_bump_time: Uint64;
+};
+
+export type ExtendRef = {
+  self: AccountAddress;
+};
