+++ src/typescript/api/src/types/index.ts
@@ -1,0 +1,2 @@
+export * from "./contract";
+export * from "./core";
